Move Dimmer Flags to a Functions Class
